+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- Ability to set height for `Markdown` field input via `rows` method.
 
 ## [1.5.0] - 2020-09-14
 ### Added